+++ resolwe/flow/tests/test_redis_cache.py
@@ -38,6 +38,11 @@ class RedisCache(TestCase):
         )
         return result
 
+    def _assertBetween(self, value, min, max):
+        """Assert value is between min and max."""
+        self.assertLessEqual(min, value)
+        self.assertGreaterEqual(max, value)
+
     def test_register_data(self):
         """Test that Data plugin is registered."""
         self.assertIn("flow.data", cache_manager._plugins)
@@ -94,6 +99,50 @@ class RedisCache(TestCase):
         self.assertLess(elapsed, 0.05)
         self.assertEqual(result, {None})
 
+        # Test locking processing for a given message.
+        identifier = [(self.data1.id, "uuid_for_test")]
+        lock_result, status = cache_manager.lock(Data, identifier)[0]
+        lock_key = redis_cache._lock_key(Data, identifier[0])
+        self.assertEqual(lock_result, True)
+        self.assertEqual(status, RedisLockStatus.PROCESSING)
+        ttl = redis_cache._redis.ttl(lock_key)
+        self._assertBetween(ttl, 290, 300)
+
+        # Locking again should return False.
+        time.sleep(2)  # Make new ttl lower.
+        lock_result, status = cache_manager.lock(Data, identifier)[0]
+        self.assertEqual(lock_result, False)
+        self.assertEqual(status, RedisLockStatus.PROCESSING)
+        new_ttl = redis_cache._redis.ttl(lock_key)
+        self._assertBetween(new_ttl, 290, 300)
+        self.assertLess(new_ttl, ttl)
+
+        # Unlocking should change the status of the lock and set TTL to 1 day.
+        cache_manager.unlock(Data, identifier)
+        lock_result, status = cache_manager.lock(Data, identifier)[0]
+        self.assertEqual(lock_result, False)
+        self.assertEqual(status, RedisLockStatus.OK)
+        ttl = redis_cache._redis.ttl(lock_key)
+        self._assertBetween(ttl, 24 * 60 * 60 - 2, 24 * 60 * 60)
+
+        # Extend the lock - only processing lock can be extended.
+        result = cache_manager.extend_lock(Data, identifier[0], valid_for=5)
+        self.assertEqual(result, False)
+        ttl = redis_cache._redis.ttl(lock_key)
+        self._assertBetween(ttl, 24 * 60 * 60 - 2, 24 * 60 * 60)
+        redis_cache._redis.unlink(lock_key)
+
+        # Extend the lock.
+        lock_result, status = cache_manager.lock(Data, identifier)[0]
+        lock_key = redis_cache._lock_key(Data, identifier[0])
+        self.assertEqual(lock_result, True)
+        self.assertEqual(status, RedisLockStatus.PROCESSING)
+        ttl = redis_cache._redis.ttl(lock_key)
+        self._assertBetween(ttl, 290, 300)
+        result = cache_manager.extend_lock(Data, identifier[0], valid_for=10)
+        ttl = redis_cache._redis.ttl(lock_key)
+        self._assertBetween(ttl, 9, 10)
+
         # Lock the data1 and wait for it.
         cache_manager.lock(Data, [(self.data1.id,)])
         start = time.time()
@@ -132,6 +181,10 @@ class RedisCache(TestCase):
         self.assertEqual(result, {RedisLockStatus.OK})
 
         # Try the lock unlocks itself after timeout.
+        lock_key = redis_cache._lock_key(Data, (self.data1.id,))
+        redis_cache._redis.unlink(lock_key)
+        lock_key = redis_cache._lock_key(Data, (self.data2.id,))
+        redis_cache._redis.unlink(lock_key)
         redis_cache.lock(Data, identifiers_list, valid_for=1)
         start = time.time()
         result = redis_cache.wait(